Make the prison system explicit
I considered using `Location.External` for non-probation systems and
marking everything else `Location.Internal` -- as this would mark the
enterprise boundary in every diagram.

("Enterprise" here is "Probation in HMPPS")

However, it turned out I cannot easily determine what is inside or
outside the enterprise and a lot of diagrams had massive black boxes in
them as everything was "inside".

So instead of the above, I opted to tag the prison systems
`prison-service` and colour them yellow (to contrast the blue).

I also realised that the NOMIS Data Hub is already represented by `NDH`
technology on the relationships, and there was no need for it to exist
as a system -- at this level. Perhaps we should style those
relationships differently?