Make probation roles accessible from model objects
People roles are generally shared across many systems; it _feels_ it's
safe to extract those roles into a model object.

This change helps with further diluting the `probation/model.kt` file.

The only weird thing this introduces is the need to mention
`ProbationPractitioners` in the prison workspace, otherwise the runtime
fails because `Delius` model in the prison workspace cannot dereference
these roles.