Represent API docs links in the model
The `APIDocs` class will set the Structurizr property called
`api-docs-url`.

The intent is to capture the "default" API docs URLs for C4 containers
that are exposing HMPPS-usable APIs. Later, it may be possible to pull
this information based on GitHub repository links, but for now, manual
fill seems appropriate.

The existence of this also allows for people to _search_ for `API` and
get results within the codebase -- if they'd like to find out what APIs
exist.