Model relationships enabled by NDH
NDH has the characteristics of a 'hub'-like system. It either pulls from
or pushes to OASys, nDelius and NOMIS.

These relationships make sense, however, if we want to have high-level
view, this communication pattern creates noise.

If we assume NDH to be a "protocol" instead of "system" for the purposes
of high-level views, we can map out the relationships enabled by the
system (`defineImpliedRelationships()`).

For more detailed views, we can keep the implementation details.