Only show NOMIS authentication flow in OMiC
The service is only used by prison staff.

HMPPS Auth has an option to use Delius as an identity provider, but that
is irrelevant for this service.

Since the NOMIS system is not on the diagram explicitly, the only way to
display the NOMIS identity provider is to add the container in NOMIS
which represents that relationship.

None of these are obvious.

It would be better if we could use `HMPPSAuth.prisonIdentity` and
`HMPPSAuth.communityIdentity` or similar to avoid untangling the
additional relationships by hand.