Transfer ownership of decisions to the owner team
Having ADRs specific to a software system shared here has its benefits.
One of the drawbacks is notifying _all_ reviewers when they change.

That model may work if we have processes around reviewing decisions at
this level, but for now it's noise for the rest of the owners.

For now, only notify/require reviews on system-specific decisions from
the system owner team.